A pointer move is a pointer move again
The precondition the five doors shared existed for one kind: the registry name
inside the version about to become current had to be free. Servers do not have
versions in the vault any more, so nothing is left to check first — each door
calls the custody move directly and branches on exactly the outcome it gets.
Unarchiving loses the same claim, and with it the two refusals it could produce.

The browser twins keep their suite, now proving what still holds there: an
approve moves the pointer, a roll-back carries a version forward, and the vault's
own answer for something it does not hold comes back in its own words.
